requirements.txt
PyQt5>=5.15.0
pyautogui>=0.9.52
pynput>=1.7.0
numpy>=1.19.0
Pillow==10.0.0 